Fix data type mismatch by casting types

+++ src/arcface/arcface_utils.py
@@ -33,7 +33,7 @@ class ArcMarginPenaltyLogits(tf.keras.layers.Layer):
         normed_w = tf.nn.l2_normalize(self.w, axis=0, name='normed_weights')
 
         cos_t = tf.matmul(normed_embds, normed_w, name='cos_t')
-        sin_t = tf.sqrt(1. - cos_t ** 2, name='sin_t')
+        sin_t = tf.cast(tf.sqrt(1. - cos_t ** 2, name='sin_t'), cos_t.dtype)
 
         cos_mt = tf.subtract(
             cos_t * self.cos_m, sin_t * self.sin_m, name='cos_mt')
